semantic_search/requirements.txt
sentence-transformers>=2.2.2
pinecone>=5.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
pytest>=7.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
Werkzeug>=3.0.0
mangum>=0.17.0
